Refs #28161 - Bookmarks light refactoring   fix ForemanForm issue
