+++ webpack/assets/javascripts/react_app/components/Bookmarks/Bookmarks.js
@@ -9,76 +9,75 @@ import { STATUS } from '../../constants';
 import { noop } from '../../common/helpers';
 import { sprintf, translate as __ } from '../../../react_app/common/I18n';
 
-class Bookmarks extends React.Component {
-  loadBookmarks = () => {
-    const { bookmarks, status, url, controller, getBookmarks } = this.props;
+const Bookmarks = props => {
+  const loadBookmarks = () => {
+    const { bookmarks, status, url, controller, getBookmarks } = props;
 
     if (bookmarks.length === 0 && status !== STATUS.PENDING) {
       getBookmarks(url, controller);
     }
   };
 
-  render() {
-    const {
-      controller,
-      url,
-      canCreate,
-      bookmarks,
-      errors,
-      status,
-      documentationUrl,
-      onBookmarkClick,
-      setModalOpen,
-    } = this.props;
+  const {
+    controller,
+    url,
+    canCreate,
+    bookmarks,
+    errors,
+    status,
+    documentationUrl,
+    onBookmarkClick,
+    setModalOpen,
+    setModalClosed,
+  } = props;
 
-    return (
-      <React.Fragment>
-        <SearchModal controller={controller} url={url} />
-        <Dropdown pullRight id={controller} onClick={this.loadBookmarks}>
-          <Dropdown.Toggle title={__('Bookmarks')}>
-            <Icon type="fa" name="bookmark" />
-          </Dropdown.Toggle>
-          <Dropdown.Menu className="scrollable-dropdown">
-            {canCreate && (
-              <MenuItem
-                key="newBookmark"
-                id="newBookmark"
-                onClick={setModalOpen}
-              >
-                {__('Bookmark this search')}
-              </MenuItem>
-            )}
-            <DocumentationUrl href={documentationUrl} />
-            <MenuItem divider />
-            <MenuItem header>{__('Saved Bookmarks')}</MenuItem>
-            {status === STATUS.PENDING && (
-              <li className="loader-root">
-                <Spinner size="xs" loading />
-              </li>
-            )}
-            {status === STATUS.RESOLVED &&
-              ((bookmarks.length > 0 &&
-                bookmarks.map(({ name, query }) => (
-                  <Bookmark
-                    key={name}
-                    text={name}
-                    query={query}
-                    onClick={onBookmarkClick}
-                  />
-                ))) || <MenuItem disabled> {__('None found')}</MenuItem>)}
-            {status === STATUS.ERROR && (
-              <MenuItem key="bookmarks-errors">
-                <EllipisWithTooltip>
-                  {sprintf('Failed to load bookmarks: %s', errors)}
-                </EllipisWithTooltip>
-              </MenuItem>
-            )}
-          </Dropdown.Menu>
-        </Dropdown>
-      </React.Fragment>
-    );
-  }
-}
+  return (
+    <React.Fragment>
+      <SearchModal
+        controller={controller}
+        url={url}
+        setModalClosed={setModalClosed}
+      />
+      <Dropdown pullRight id={controller} onClick={loadBookmarks}>
+        <Dropdown.Toggle title={__('Bookmarks')}>
+          <Icon type="fa" name="bookmark" />
+        </Dropdown.Toggle>
+        <Dropdown.Menu className="scrollable-dropdown">
+          {canCreate && (
+            <MenuItem key="newBookmark" id="newBookmark" onClick={setModalOpen}>
+              {__('Bookmark this search')}
+            </MenuItem>
+          )}
+          <DocumentationUrl href={documentationUrl} />
+          <MenuItem divider />
+          <MenuItem header>{__('Saved Bookmarks')}</MenuItem>
+          {status === STATUS.PENDING && (
+            <li className="loader-root">
+              <Spinner size="xs" loading />
+            </li>
+          )}
+          {status === STATUS.RESOLVED &&
+            ((bookmarks.length > 0 &&
+              bookmarks.map(({ name, query }) => (
+                <Bookmark
+                  key={name}
+                  text={name}
+                  query={query}
+                  onClick={onBookmarkClick}
+                />
+              ))) || <MenuItem disabled> {__('None found')}</MenuItem>)}
+          {status === STATUS.ERROR && (
+            <MenuItem key="bookmarks-errors">
+              <EllipisWithTooltip>
+                {sprintf('Failed to load bookmarks: %s', errors)}
+              </EllipisWithTooltip>
+            </MenuItem>
+          )}
+        </Dropdown.Menu>
+      </Dropdown>
+    </React.Fragment>
+  );
+};
 
 Bookmarks.propTypes = {
   controller: PropTypes.string.isRequired,
@@ -91,6 +90,7 @@ Bookmarks.propTypes = {
   documentationUrl: PropTypes.string,
   getBookmarks: PropTypes.func,
   setModalOpen: PropTypes.func.isRequired,
+  setModalClosed: PropTypes.func.isRequired,
 };
 
 Bookmarks.defaultProps = {
